DRIVERS: SLIMBUS: some more tweaks:
Set ROOT_FREQ at 22, which is double what we had before. According to some documents this is
what we want for optimium audio quality
Forcec SLIM_RATE to 44.1KHz and 48KHz instead of much lower values. This way it will be
forced to run at these freqs to match the majority of audio files user will be playing.

ZeroInfinity/Poondog

Signed-off-by: ZeroInfinity <infinitezero345@gmail.com>
Signed-off-by: poondog <markj338@gmail.com>